+++ src/client/widgets/scripts/script-vm.js
@@ -326,8 +326,28 @@ class ScriptVm extends Vm {
 
         }
 
+
+        this.sandbox.contentWindow.setFocus = (id)=>{
+
+            if (id) {
+                var widgets = this.resolveId(id)
+
+                if (widgets.length) {
+                    setTimeout(()=>{
+
+                        widgets[0].widget.focus()
+                    })
+                }
+            }
+
+            // built-in client only: electron will call window.focus()
+            console.debug('ELECTRON.FOCUS()')
+
+        }
+
         this.sandbox.contentWindow.unfocus = ()=>{
 
+            window.blur()
             // built-in client only: electron will call window.blur()
             console.debug('ELECTRON.BLUR()')
 
@@ -354,7 +374,7 @@ class ScriptVm extends Vm {
         }
 
         for (var imports of ['set', 'get', 'getProp', 'getIndex', 'updateProp', 'send', 'httpGet', 'stateGet', 'stateSet', 'storage',
-            'setInterval', 'clearInterval', 'setTimeout', 'clearTimeout', 'unfocus', 'setScroll', 'getScroll', 'toolbar']) {
+            'setInterval', 'clearInterval', 'setTimeout', 'clearTimeout', 'setFocus', 'unfocus', 'setScroll', 'getScroll', 'toolbar']) {
             this.sanitize(this.sandbox.contentWindow[imports])
         }